Week 8 Check-In
Okay. That was really weird. For some reason, my changes weren’t
showing in GitHub, so I deleted and then reloaded the IDM folder to it.
I wonder if this is because I change the parent folder IDM was in from
Responsive Design (with a space) to Responsive Design (no space), which
I had to do in order for Terminal to be able to find the directory. I
wonder if this is why when I came into GitHub it read “Head” instead of
“Master”… super weird. Hope I didn’t break anything…